Update dremio-setup.md
The sentence 
"See Support Keys in the Dremio documentation for the steps." makes no sense in this context. There is nothing on this page that requires the setting of a support key, apart from if you're using a PAT and haven't yet set up your Dremio system to use PATs, and even then, the wrong page is linked to.